requirements.txt
pipdeptree==0.13.2
pytest-cov==2.8.1
pytest-mock==2.0.0
requests-mock==1.7.0
setuptools==41.2.0
